Separate lifting into its own module
